Move around hardware comm files
