+++ scratch_work/test.py
@@ -7,4 +7,6 @@ motor = kinesis.KinesisMotor(devices[0][0], scale="stage")
 # print(motor.get_scale_units())
 # print(motor.get_scale())
 # print(motor.get_stage())
-print(motor.get_limit_switch_parameters())
+#print(motor.get_limit_switch_parameters(scale=True))
+print(motor.get_full_info())
+motor.move_by(distance=-1)
